backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1

# gRPC Dependencies
grpcio>=1.60.0
grpcio-tools>=1.60.0
grpcio-status>=1.60.0

# RAG Dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-pinecone>=0.0.1
langchain-google-genai>=0.0.6
pinecone-client>=2.2.4
numpy>=1.24.0
pdfplumber>=0.10.0